backend/requirements.txt
# Educational Vulnerability Scanner - Backend Dependencies
# Python 3.11+ required

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Network Scanning
python-nmap==0.7.1

# AI Integration
openai==1.3.5

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1

# Development/Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
